Spring junit test added

+++ src/main/java/by/makei/spring/service/UserService.java
@@ -11,5 +11,5 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final CrudRepository<Integer, Company> companyRepository;
+    private final CrudRepository<Long, Company> companyRepository;
 }
